REFACTOR: Not using folders as there is only 1 item per folder
