+++ app/auth/2fa/page.tsx
@@ -4,8 +4,8 @@ import { redirect } from "next/navigation";
 import { headers } from "next/headers";
 import { auth } from "@/lib/auth/auth";
 import { ROUTES } from "@/lib/routes";
-import { TotpForm } from "./_components/forms/totp-form";
-import { BackupCodeTab } from "./_components/tabs/backup-code-tab";
+import { TotpForm } from "./_components/totp-form";
+import { BackupCodeTab } from "./_components/backup-code-tab";
 
 const TAB_VALUES = {
   TOTP: "totp",
